requirements.txt
streamlit
git+https://github.com/JaidedAI/EasyOCR.git
opencv-python-headless
numpy
Pillow
requests
transformers
groq
torch
torchvision>=0.5
opencv-python-headless
scipy
fuzzywuzzy
scikit-image
python-bidi
PyYAML
Shapely
pyclipper
ninja
# paddleocr
# paddlepaddle-gpu